Suggested changes.
