+++ config/sync/group.role.personal-administrator.yml
@@ -211,3 +211,4 @@ permissions:
   196: 'administer apps'
   197: 'administer group'
   198: manage_menu
+  394: 'download citations'
