+++ config/sync/group.role.personal-anonymous.yml
@@ -12,6 +12,7 @@ audience: anonymous
 group_type: personal
 permissions_ui: true
 permissions:
-  - 'view group_node:person entity'
-  - 'view group_entity:bibcite_reference entity'
-  - 'view group_entity:bibcite_reference content'
+  0: 'view group_node:person entity'
+  1: 'view group_entity:bibcite_reference entity'
+  2: 'view group_entity:bibcite_reference content'
+  6: 'download citations'
