+++ config/sync/group.role.personal-outsider.yml
@@ -12,7 +12,8 @@ audience: outsider
 group_type: personal
 permissions_ui: true
 permissions:
-  - 'view group'
-  - 'view group_node:person entity'
-  - 'view group_entity:bibcite_reference entity'
-  - 'view group_entity:bibcite_reference content'
+  0: 'view group'
+  1: 'view group_node:person entity'
+  2: 'view group_entity:bibcite_reference entity'
+  3: 'view group_entity:bibcite_reference content'
+  8: 'download citations'
