+++ profile/modules/apps/os_publications/os_publications.group.permissions.yml
@@ -1,0 +1,4 @@
+download citations:
+  title: 'Donwload citations'
+  description: 'Allow users to export/download citations.'
+  restrict access: false
