+++ profile/modules/apps/os_publications/os_publications.module
@@ -491,7 +491,8 @@ function os_publications_preprocess_bibcite_citation(&$variables) {
 
     // Get export button render array , get it's html and append to citation.
     $export_array = \Drupal::service('os_publications.citation_helper')->getCitationDownloadButton($id);
-    if ($export_array) {
+    $vsite = \Drupal::service('vsite.context_manager')->getActiveVsite();
+    if ($export_array && $vsite) {
       $export_button = \Drupal::service('renderer')->render($export_array);
       $variables['content']['#markup'] .= $export_button;
       $variables['content']['#allowed_tags'] = [
